Prince+

+++ app/src/main/java/dev/groupx/apkikala/ui/screen/sign_up/UploadProfileImageViewModel.kt
@@ -1,13 +1,13 @@
 package dev.groupx.apkikala.ui.screen.sign_up
 
 import android.net.Uri
+import android.util.Log
 import androidx.compose.runtime.mutableStateOf
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dev.groupx.apkikala.model.service.AccountService
 import dev.groupx.apkikala.model.service.LogService
 import dev.groupx.apkikala.model.service.StorageService
 import dev.groupx.apkikala.ui.screen.ApkiKalaViewModel
-import dev.groupx.apkikala.ui.screen.home.HomeNode
 import javax.inject.Inject
 
 @HiltViewModel
@@ -28,24 +28,27 @@ class UploadProfileImageViewModel @Inject constructor(
         }
     }
 
-    fun onAddProfileImageClick(restartApp: (String) -> Unit) {
+    fun onAddProfileImageClick(restartApp: (String) -> Unit, destRoute: String) {
         launchCatching {
             storageService.saveImageToFirestoreUser(setUrl, accountService.currentUserId)
-            restartApp(HomeNode.route)
+            restartApp(destRoute)
         }
     }
 
-    fun onCancel(popUp: () -> Unit) {
+    fun onCancel(popUp: () -> Unit, destRoute: String) {
         launchCatching {
             storageService.removeImage()
-            accountService.deleteAccount(accountService.currentUserId)
+            if (destRoute == "${UploadProfileImageNode}/HomeScreen")
+                accountService.deleteAccount(accountService.currentUserId)
             popUp()
         }
     }
 
-    fun onBackClick(popUp: () -> Unit) {
+    fun onBackClick(popUp: () -> Unit, destRoute: String) {
         launchCatching {
-            accountService.deleteAccount(accountService.currentUserId)
+            Log.d("In UploadProfile", destRoute)
+            if (destRoute == "${UploadProfileImageNode}/HomeScreen")
+                accountService.deleteAccount(accountService.currentUserId)
             popUp()
         }
     }
